default reports are no longer separate executables
this is to facilitate installation via homebrew, as it
_really_ only wants you to install one executable file and doesn't
give you access to the users home dir.

this also makes it harder for users to accidentally hose the
bundled reports.

user created reports with the same name as a bundled report
will replace that bundled report.

user created reports are still expected to conform to the same
arguments and json responses and must be placed in
~/.config/hey/reports/