Events now created with Time.utc_now instead of Time.now
Time.now is documented as UTC but in reality it's localtime. There may be some
UTC backing to it but when you convert it to a string you get
localtime not UTC time so... *shrug*

This change means event created_at dates are now stored in UTC in SQLite
which means they're now correctly extracted as localtime.